app/tools/threat_intel_platform/requirements.txt
Flask>=2.0.0
Flask-SQLAlchemy>=3.0.0
Flask-Cors>=3.0.10
requests>=2.28.0
beautifulsoup4>=4.11.1
python-dotenv>=0.20.0
APScheduler>=3.9.0
OTXv2==1.5.12
pymisp==2.4.170
cabby==0.1.23
taxii2-client>=2.3.0
stix2>=3.0.0
stix2-patterns==2.0.0
flask-restful==0.3.9
flask-jwt-extended==4.4.4
praw==7.7.0
scrapy==2.8.0
celery==5.2.7
redis==4.5.1
regex==2023.3.23
gunicorn==20.1.0
sqlalchemy>=2.0.0
marshmallow==3.19.0
pytz>=2022.1
pyyaml>=6.0
cryptography>=38.0.0
python-dateutil>=2.8.2
urllib3>=1.26.12
chardet>=5.0.0
lxml>=4.9.1
defusedxml>=0.7.1
validators>=0.20.0 